[posix] fix the snprintf buffer overflow issue
Based on the description of the function [snprintf](https://linux.die.net/man/3/vsnprintf),
the return value of size or more means that the output was truncated.
If the return value is larger than the buffer size and the code access
the buffer using the return value as the buffer length, it causes the
buffer overflow.