[spinel] Add iid-list option to radio_url.
List of IIDs a host can subscribe to receive spinel frames other than provided in 'iid' argument. If not specified, host will subscribe to the interface ID provided in 'iid` argument. Valid values are 0-3.

Also, remove the use of openthread config option from spinel.h as a part of the review comment.