[routing-manager] support Advertising PIO (AP) flag in published route (#9274)
This commit adds support for the Advertising PIO (AP) flag in Network
Data `ExternalRoute` entries. It also updates the `RoutingManager`
class to set the AP flag on its published Network Data route when the
local on-link prefix is included as PIO in emitted RA messages, which
is equivalent to when the local on-link prefix is being published,
advertised, or deprecated.

This commit also updates the `test_routing_manager` unit test to
validate the AP flag in the published route under different
scenarios.