[posix] replace RCP vendor extension find package with include command (#9241)
This change updates the mechanism used by vendor.cmake to include a
cmake file to define the required vendor targets used by the rcp
vendor interface library instead of the find_package command that was
previously used.

The find_package command would allow for a package to be searched for
and a target to be linked unconditionally if defined by the CMake
project. This would lead to problems if the target was already
defined, perhaps as a subdirectory of a super project, and the cache
variable is still set for the openthread build. Additionally, this
change also brings the vendor extension of the posix library more in
line with the cli vendor extension, which was created after
find_package was used in vendor.cmake.