[mle] change `protected` declarations as `private
This commit changes the `protected` declarations in the `Mle` class to
be `private` since `Mle` class and its subclass, the `MleRouter`, are
declared as `friend` classes.

The commit also re-arranges the `private` definitions in the `Mle`
class into five sections: constants, enumerations, nested classes,
methods declarations, and member variable definitions.

With the change to `private`, this commit also removes the redundant
documentation of methods.