[mesh-forwarder] prevent double freeing of aged message
When TX message queue management was enabled, it was observed an attempt
to dequeue and free the message, that was being processed at the time,
has failed and led to an assert.
The message's buffer was previously freed while removing aged messages.

Add callback to `RemoveAgedMessages` method so freeing of currently
processed message can be detected and further attempt to free it can
be skipped.