[dsnsd-server] support handling of "A record" queries
This commit adds support for responding to "A record" queries in the
DNS-SD server and the discovery proxy.

If the query is resolved by SRP (i.e., the hostname matches a host on
the SRP server), the host's IPv6 addresses are returned in the
Additional Data Section. When resolved by proxy, the `otPlatDnssd`
APIs are used to start/stop IPv4 address resolvers for the name on
the infrastructure network.

This commit also adds helper methods to detect if an IPv6 address
follows the IPv4-mapped IPv6 address format using the `::ffff:0:0/96`
prefix (e.g., `::ffff:192.0.2.128`) and parse or or construct such
addresses.

Unit test `test_dnssd_discovery_proxy` and `test_ip_address` are also
updated to check the newly added functionalities.