[mle] allow retries of MLE Child ID Request message
This commit adds a retry mechanism in MLE layer for sending Child ID
Requests during an attach cycle. The number of attempts is determined
by the constant `kChildIdRequestAttempts`, which is set to two. The
retries occur evenly spaced during the `kChildIdResponseTimeout`
interval (1250 msec). So, with two attempts, the MLE layer waits 625
msec to retransmit after the first transmission, and then waits
another 625 msec for a total wait time of 1250 msec to receive the
Child ID Response.