[Mac] get the supported channel mask directly from platform API
The current code of the `mac.cpp` caches the supported channel mask to a
local variable. But the supported channel mask may be changed after the
country code is changed. This will cause the supported channel mask used
in `mac.cpp` to be inconsistent with the actual supported channel mask.

This commit makes `mac.cpp` to get the supported channel mask directly
from the platform API to avoid the channel mask inconsistencies.