[mac] document and adjust to standards based timing concepts (#9322)
Adds a precise specification for the local radio clock that is at the
core of all timing-sensitive aspects of the IEEE 802.15.4 protocol
(currently mostly CSL) and refers to this definition for fields that
reference the radio clock.

Specifies the reference planes and message timestamp points for RX
timestamps, RX windows and timed TX.

Documents the Thread-specific interpretation of the CSL Phase and
derives the formula for CSL TX timestamps from it.

Based on the standard based definitions given, the TX timestamp used
for timed TX now refers to the start of PHR. This change needs to be
synchronized with the platform radio driver implementations. An
appropriate change set has been prepared for Zephyr.