[mle] simplify reattach on losing connectivity to leader (#9479)
This commit updates the attach process to the same partition after
a router/REED losing connectivity to the leader. It removes the
`kSamePartitionRetry` enumerator from `AttachMode`. With this change,
after trying to attach using `kSamePartition` attach mode (with a
total of six Parent Request attempts to Routers and REEDs), the
device will switch to `kAnyPartition` attach mode. This helps to
reduce the total number of attempts before the device can decide to
take the leader role.