[ping-sender] allow setting multicast loop (#9494)
This commit updates `PingSender` and its config to include a
`mMulticastLoop` flag and allow the caller to set this flag. When
set, multicast ping echo request messages are looped back and
received by the device itself if it is subscribed to the same
address. The `ping` CLI command is also updated to allow setting this
flag.

This commit also updates `test-008-multicast-traffic.py` to add new
test cases to cover the multicast loop flag behavior (pinging with or
without this flag). In particular, it covers when the ping
destination is a realm-local multicast address and when it is a
larger-than-realm-local multicast address (where the `ot::Ip6` module
would use a tunnel header for forwarding the message).