[ip6-mpl] set `LoopbackToHostAllowed` to false for MPL retx
This commit updates `Mpl::HandleRetransmissionTimer()` to set
`SetLoopbackToHostAllowed(false)` on MPL message retransmissions.

If a multicast message originates from the host itself, it can be
passed to the host when the original message is processed. If it
originates from the Thread Netif (received over Thread radio), it can
be passed to the host when the received message is processed. In both
cases, we no longer need to pass the MPL retransmissions of the
message to the host.