update the hal_interface to fix the cherry-pick issue
OpenThread has changed the functions of SpinelInterface and
changed the code to support both the HDLC and SPI interfaces at
the same time, it causes the build errors of the OpenThread in
Android. This commit updates the hal_interface based on the
latest definition of SpinelInterface and updates the spinel
interface configurations.

Bug: b/304914440
Test: Build and run Thread stack on cuttlefish.
Change-Id: Ic4e4be0e268e3052244778bab9df4fbbdd01aefd