[radio-spinel] simplify `SetMacKey()` (#9539)
This commit simplifies the `RadioSpinel::SetMacKey()` method.
A common `private` overload of `SetMacKey()` is added that uses
`const otMacKey &` as its input key type.

When `KEY_REFERENCES_ENABLE` is not enable the `SetMacKey()` simply
passes the key from the `otMacKeyMaterial` directly. This avoids the
unnecessary copying of the key into a local variable.

When `KEY_REFERENCES_ENABLE` is enabled, the keys are read from
secure storage. A new `ReadMacKey()` helper method is added which
also validates that the read key size is correct.